Remove counter related code, rename state folder
